requirements.txt
flask
flask-uploads
opencv-python
matplotlib
pandas
numpy
dash
plotly.express
Pillow
openapi-codec